Remove utmpset program
The utmpset program was written in 2002, updated in 2003 to support
solaris, and since then only got few bugfixes. Over time it broke on
most systems, and never was fail-safe.  Today, portability of utmp[x]
and even more so wtmp[x] remains tricky. Many packages have tried or
done this already, or reimplemented utmp[x]/wtmp[x].

Let's remove and not fix.  Ideally any utmp[x]/wtmp[x] implementation
would include a program to edit utmp[x] and append to wtmp[x] database,
but it is not so.